Last significant refactor, I promise. Abandoning Histogram class altogether since it didn't do anything special and most users will be interfacing with ROOT, so we use TH1 directly.
Most math behind the scenes is still done with Eigen to try to reduce the number of temporaries through the use of Eigen::Ref and expressions.

Provide a bunch of utility functions for ROOT<->Eigen conversions.

Estimators (still IMeasurement) objects use IMeasurement facilities to convert ROOT histograms to Eigen arrays and dispatch evaluation of the estimation to the derived class through a private virtual function.

Most everything now builds into a library. Still need to update most of the tests.